requirements.txt
python-telegram-bot
llama-index >= 0.6
python-dotenv == 1.0.0